pyproject.toml
[project]
name = "agno-cookbook"
version = "0.1.0"
description = "Agno Cookbook 中文版 - AI Agent 开发示例与教程"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agno",
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "yfinance>=1.1.0",
]

[dependency-groups]
quickstart = [
    "chromadb",
    "fastapi[standard]",
    "google-genai",
    "mcp",
    "nest-asyncio",
    "openinference-instrumentation-agno",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pgvector",
    "pillow",
    "psycopg[binary]",
    "sqlalchemy",
    "yfinance",
]
demo = [
    "anthropic",
    "duckduckgo-search",
    "duckdb",
    "exa-py",
    "openai",
    "pgvector",
    "psycopg[binary]",
    "sqlalchemy",
    "yfinance",
]
learning = [
    "anthropic",
    "chromadb",
    "fastapi[standard]",
    "google-genai",
    "mcp",
    "nest-asyncio",
    "openai",
    "openinference-instrumentation-agno",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "pgvector",
    "pillow",
    "psycopg[binary]",
    "sqlalchemy",
]
mcp-toolbox = [
    "fastapi",
    "mcp",
    "openai",
    "sqlalchemy",
    "toolbox-core",
]
all = [
    {include-group = "quickstart"},
    {include-group = "demo"},
    {include-group = "learning"},
    {include-group = "mcp-toolbox"},
]
